refacto functions and tests on shadow meeting
