+++ web/misc/delete_old_shadow_meetings.py
@@ -1,3 +1,3 @@
-from b3desk.models.meetings import delete_all_shadow_meetings
+from web.b3desk.models.meetings import delete_all_old_shadow_meetings
 
-delete_all_shadow_meetings()
+delete_all_old_shadow_meetings()
